Updated routing
1. Relocated a routing array to the app-routing.module
2. Added directives routerLink & routerLinkActive
3. Fixed the MobMenu closing after navigate on a new url
