+++ frontend/src/app/app-routing.module.ts
@@ -1,10 +1,20 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { MainComponent } from './components/main/main.component';
+import { RaftingComponent } from './components/rafting/rafting.component';
+import { SemeyniyRaftingComponent } from './components/semeyniy-rafting/semeyniy-rafting.component';
+import { Error404Component } from './components/error404/error404.component';
 
-const routes: Routes = [];
+// Роутинг
+const appRoutes: Routes = [
+  {path: '', component: MainComponent},
+  {path: 'rafting', component: RaftingComponent},
+  {path: 'semeyniy-rafting', component: SemeyniyRaftingComponent},
+  {path: '**', component: Error404Component}
+];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [RouterModule.forRoot(appRoutes)],
   exports: [RouterModule]
 })
 export class AppRoutingModule { }
