+++ frontend/src/app/components/first-block/first-block.component.html
@@ -1,7 +1,7 @@
 <div class="container container-main">
 	  
     <div class="logo-mobile">
-       <a href="#"><img src="../../assets/img/logo.png" title="База рафтинга в Лосево «Кивиниеми»" alt="База рафтинга в Лосево «Кивиниеми»"></a>
+       <a [routerLink]="['/']"><img src="../../assets/img/logo.png" title="База рафтинга в Лосево «Кивиниеми»" alt="База рафтинга в Лосево «Кивиниеми»"></a>
     </div>
 
     <!-- Рафтинг -->
@@ -24,7 +24,7 @@
             <div class="item">
                     <h3 class="item-title">Рафтинг</h3>
                     <div class="primary-wrapper">
-                        <a href="/rafting" class="btn-more" title="Рафтинг в Лосево, почувствуй вкус лета!">
+                        <a [routerLink]="['/rafting']" class="btn-more" title="Рафтинг в Лосево, почувствуй вкус лета!">
                             <span class="btn-more-active"></span>
                             <span class="btn-more-inner" id="rafting-item">Подробнее</span>
                         </a>
