+++ frontend/src/app/components/rafting/rafting.component.html
@@ -1,7 +1,7 @@
 <div class="container container-main">	  
         
     <div class="logo-mobile">
-        <a href="/"><img src="../../assets/img/logo.png" title="База рафтинга в Лосево «Кивиниеми»" alt="База рафтинга в Лосево «Кивиниеми»"></a>
+        <a [routerLink]="['/']"><img src="../../assets/img/logo.png" title="База рафтинга в Лосево «Кивиниеми»" alt="База рафтинга в Лосево «Кивиниеми»"></a>
     </div>
  
     <!-- Рафтинг -->
@@ -15,7 +15,7 @@
                 <div class="item">
                      <h3 class="item-title">Семейный рафтинг</h3>
                      <div class="primary-wrapper">
-                         <a href="/semeyniy-rafting" class="btn-more" title="Семейный рафтинг - катаем детей от 7-ми лет, аквапарк на природе!">
+                         <a [routerLink]="['/semeyniy-rafting']" class="btn-more" title="Семейный рафтинг - катаем детей от 7-ми лет, аквапарк на природе!">
                              <span class="btn-more-active"></span>
                              <span class="btn-more-inner">Подробнее</span>
                          </a>
